Move discard schools script to data hub module
